requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
joblib
pyyaml
fastapi
uvicorn[standard]
openpyxl
matplotlib
pytest
